refactor(user): GroupPlugins trait, Arc<dyn Identity>, split user/mod.rs
